work around issue for verification of testcases
it's not straightforward to verify a result if the only thing you want
to do is just syntatical verification, because the textual output of a form
doesn't read back in to the same form.

So is(eq=parse_string(string(eq))) is not always true. An insightful
thread is available on the maxima list, subject is "question about
verifying maxima results".